refactor: group src/ into views/ widgets/ platform/ (tiered layout)
- src/views/    the 16 feature modules
- src/widgets/  gpui-coupled widgets (editor, mdview, terminal, remote_img)
- src/platform/ small OS utils (clipboard, scratch, shellcmd)
Top level keeps the core shell (main/app/render) + shared infra (ui/, divider).
Widgets/platform re-aliased in main.rs so editor::/mdview::/scratch:: paths are
unchanged; views' use super::* → use crate::*. Zero behaviour change.
