+++ src/platform/mod.rs
@@ -1,0 +1,4 @@
+//! Small OS-integration utilities (no gpui): clipboard, shell-command install, scratch paths.
+pub(crate) mod clipboard;
+pub(crate) mod scratch;
+pub(crate) mod shellcmd;
