+++ src/platform/scratch.rs
